+++ kythe/javatests/com/google/devtools/kythe/analyzers/java/testdata/pkg/Comments.java
@@ -14,7 +14,8 @@ package pkg;
  * This is a Javadoc comment with links to {@link String}, {@link java.lang.Integer}, and
  * {@link Inner}.
  */
-public class Comments {
+public class Comments
+    implements Comparable<Comments> {
 
   //- @+3"// inline comment here" documents FieldOne
   //- @+2fieldOne defines/binding FieldOne
@@ -47,6 +48,31 @@ public class Comments {
   //- @+3InnerE defines/binding InnerE
 
   /** This comments the {@link InnerE} enum. */
-  static enum InnerE {}
-  // TODO(schroederc): link JavaDoc
+  static enum InnerE {
+
+    //- @+3"// This comments SOME_VALUE." documents SomeValue
+    //- @+3SOME_VALUE defines SomeValue
+
+    // This comments SOME_VALUE.
+    SOME_VALUE,
+
+    //- @+3"/** This documents {@link ANOTHER_VALUE}. */" documents AnotherValue
+    //- @+3ANOTHER_VALUE defines AnotherValue
+
+    /** This documents {@link ANOTHER_VALUE}. */
+    ANOTHER_VALUE;
+  }
+
+  //- @+3"/** This documents {@link #toString()}. */" documents ToString
+  //- @+3toString defines/binding ToString
+
+  /** This documents {@link #toString()}. */
+  @Override public String toString() { return "null"; }
+
+  //- @+3"/** This documents {@link #compareTo()} over {@link Override}. */" documents CompareTo
+  //- @+4compareTo defines/binding CompareTo
+
+  /** This documents {@link #compareTo()} over {@link Override}. */
+  @Override
+  public int compareTo(Comments o) { return 0; }
 }